Created the util Scala backend project and moved PythonTranslator and ProtoBufCodec into it
