+++ geopyspark-backend/build.sbt
@@ -33,6 +33,10 @@ scalaVersion := Version.scala
 
 lazy val root = Project("root", file("."))
 
-lazy val geotrellisProject = Project("geotrellis-backend", file("geotrellis"))
+lazy val utilProject = Project("util", file("util"))
   .settings(commonSettings: _*)
   .dependsOn(root)
+
+lazy val geotrellisProject = Project("geotrellis-backend", file("geotrellis"))
+  .settings(commonSettings: _*)
+  .dependsOn(root, utilProject)
