+++ geopyspark-backend/geotrellis/src/main/scala/geopyspark/geotrellis/io/ValueReaderWrapper.scala
@@ -24,7 +24,7 @@ import java.util.ArrayList
 import scala.collection.JavaConverters._
 import scala.collection.mutable
 
-import geopyspark.geotrellis.PythonTranslator
+import geopyspark.util.PythonTranslator
 
 
 /**
