+++ geopyspark-backend/geotrellis/src/main/scala/geopyspark/geotrellis/protobufs/ProjectedExtentProtoBuf.scala
@@ -1,7 +1,7 @@
 package geopyspark.geotrellis.protobufs
 
 import geopyspark.geotrellis._
-import geopyspark.geotrellis.ProtoBufCodec
+import geopyspark.util.ProtoBufCodec
 import geotrellis.vector._
 
 import protos.extentMessages._
