+++ geopyspark-backend/geotrellis/src/main/scala/geopyspark/geotrellis/protobufs/SpatialKeyProtoBuf.scala
@@ -1,5 +1,6 @@
 package geopyspark.geotrellis.protobufs
 
+import geopyspark.util._
 import geopyspark.geotrellis._
 import geotrellis.spark._
 import protos.keyMessages._
