+++ geopyspark-backend/geotrellis/src/main/scala/geopyspark/geotrellis/protobufs/TemporalProjectedExtentProtoBuf.scala
@@ -1,7 +1,7 @@
 package geopyspark.geotrellis.protobufs
 
 import geopyspark.geotrellis._
-import geopyspark.geotrellis.ProtoBufCodec
+import geopyspark.util.ProtoBufCodec
 import geotrellis.spark._
 
 import protos.extentMessages._
