+++ geopyspark-backend/geotrellis/src/main/scala/geopyspark/geotrellis/tests/schemas/SpatialKeyWrapper.scala
@@ -1,5 +1,6 @@
 package geopyspark.geotrellis.tests.schemas
 
+import geopyspark.util._
 import geopyspark.geotrellis._
 import protos.keyMessages._
 import geopyspark.geotrellis.testkit._
