+++ geopyspark-backend/geotrellis/src/main/scala/geopyspark/geotrellis/tms/TmsRoutes.scala
@@ -1,5 +1,7 @@
 package geopyspark.geotrellis.tms
 
+import geopyspark.geotrellis._
+
 import geotrellis.raster._
 import geotrellis.raster.render._
 import geotrellis.spark._
@@ -65,7 +67,7 @@ object TMSServerRoutes {
             .retrieve(zoom, x, y)
             .map(_.map{tile =>
               if (renderer.requiresEncoding()) {
-                renderer.renderEncoded(geopyspark.geotrellis.PythonTranslator.toPython(tile))
+                renderer.renderEncoded(geopyspark.util.PythonTranslator.toPython(tile))
               } else {
                 renderer.render(tile)
               }
@@ -90,7 +92,7 @@ object TMSServerRoutes {
             .map(
               _.map(array =>
                 if (compositer.requiresEncoding()) {
-                  compositer.compositeEncoded(array.map{tile => geopyspark.geotrellis.PythonTranslator.toPython(tile)})
+                  compositer.compositeEncoded(array.map{tile => geopyspark.util.PythonTranslator.toPython(tile)})
                 } else {
                   compositer.composite(array)
                 }
