+++ geopyspark-backend/util/build.sbt
@@ -1,0 +1,28 @@
+name := "util"
+
+resolvers ++= Seq(
+  "Location Tech GeoTrellis Snapshots" at "https://repo.locationtech.org/content/repositories/geotrellis-snapshots",
+  "Location Tech GeoTrellis resleases" at "https://repo.locationtech.org/content/groups/releases",
+  Resolver.mavenLocal
+)
+
+libraryDependencies ++= Seq(
+  "org.apache.spark"            %% "spark-core"            % "2.0.0" % "provided",
+  "org.locationtech.geotrellis" %% "geotrellis-s3"         % Version.geotrellis,
+  "org.locationtech.geotrellis" %% "geotrellis-spark"      % Version.geotrellis
+)
+
+assemblyMergeStrategy in assembly := {
+  case "reference.conf" => MergeStrategy.concat
+  case "application.conf" => MergeStrategy.concat
+  case "META-INF/MANIFEST.MF" => MergeStrategy.discard
+  case "META-INF\\MANIFEST.MF" => MergeStrategy.discard
+  case "META-INF/ECLIPSEF.RSA" => MergeStrategy.discard
+  case "META-INF/ECLIPSEF.SF" => MergeStrategy.discard
+  case x if x.startsWith("META-INF/services") => MergeStrategy.concat
+  case _ => MergeStrategy.first
+}
+
+PB.targets in Compile := Seq(
+  scalapb.gen() -> (sourceManaged in Compile).value
+)
