+++ geopyspark-backend/util/src/main/scala/geopyspark/util/PythonTranslator.scala
@@ -1,4 +1,4 @@
-package geopyspark.geotrellis
+package geopyspark.util
 
 import geotrellis.raster._
